requirements.txt
python-dotenv
# slack_bolt
slack-sdk
langchain
openai
pandas
faiss-cpu
tiktoken